rebase(storage): rename `TaskStorage` to `Storage`
Move `TaskStorage` setup to `mod.rs` and rename it to `Storage`.
