+++ src/io/cli/mod.rs
@@ -4,14 +4,14 @@ mod render;
 mod resolve;
 mod run;
 
-use crate::storage::TaskStorage;
+use crate::storage::Storage;
 
 pub struct Cli {
-    storage: TaskStorage,
+    storage: Storage,
 }
 
 impl Cli {
-    pub fn new(storage: TaskStorage) -> Self {
+    pub fn new(storage: Storage) -> Self {
         Self { storage }
     }
 }
